+++ src/main/java/io/github/thebusybiscuit/slimefun4/core/services/localization/SlimefunLocalization.java
@@ -1,13 +1,17 @@
 package io.github.thebusybiscuit.slimefun4.core.services.localization;
 
-import java.util.Arrays;
 import java.util.Collection;
+import java.util.Collections;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 import java.util.function.UnaryOperator;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
+import javax.annotation.ParametersAreNonnullByDefault;
 
+import org.apache.commons.lang.Validate;
 import org.bukkit.ChatColor;
 import org.bukkit.Keyed;
 import org.bukkit.NamespacedKey;
@@ -33,9 +37,9 @@ import net.md_5.bungee.api.chat.TextComponent;
 /**
  * This is an abstract parent class of {@link LocalizationService}.
  * There is not really much more I can say besides that...
- * 
+ *
  * @author TheBusyBiscuit
- * 
+ *
  * @see LocalizationService
  *
  */
@@ -48,10 +52,10 @@ public abstract class SlimefunLocalization extends Localization implements Keyed
     /**
      * This method attempts to return the {@link Language} with the given
      * language code.
-     * 
+     *
      * @param id
      *            The language code
-     * 
+     *
      * @return A {@link Language} with the given id or null
      */
     @Nullable
@@ -59,28 +63,30 @@ public abstract class SlimefunLocalization extends Localization implements Keyed
 
     /**
      * This method returns the currently selected {@link Language} of a {@link Player}.
-     * 
+     *
      * @param p
      *            The {@link Player} to query
-     * 
+     *
      * @return The {@link Language} that was selected by the given {@link Player}
      */
+    @Nullable
     public abstract Language getLanguage(@Nonnull Player p);
 
     /**
      * This method returns the default {@link Language} of this {@link Server}
-     * 
+     *
      * @return The default {@link Language}
      */
+    @Nullable
     public abstract Language getDefaultLanguage();
 
     /**
      * This returns whether a {@link Language} with the given id exists within
      * the project resources.
-     * 
+     *
      * @param id
      *            The {@link Language} id
-     * 
+     *
      * @return Whether the project contains a {@link Language} with that id
      */
     protected abstract boolean hasLanguage(@Nonnull String id);
@@ -88,7 +94,7 @@ public abstract class SlimefunLocalization extends Localization implements Keyed
     /**
      * This method returns a full {@link Collection} of every {@link Language} that was
      * found.
-     * 
+     *
      * @return A {@link Collection} that contains every installed {@link Language}
      */
     @Nonnull
@@ -96,7 +102,7 @@ public abstract class SlimefunLocalization extends Localization implements Keyed
 
     /**
      * This method adds a new {@link Language} with the given id and texture.
-     * 
+     *
      * @param id
      *            The {@link Language} id
      * @param texture
@@ -117,94 +123,154 @@ public abstract class SlimefunLocalization extends Localization implements Keyed
         }
     }
 
-    public String getMessage(Player p, String key) {
+    @Nonnull
+    private final FileConfiguration getFallback(@Nonnull LanguageFile file) {
+        Language language = getLanguage(SupportedLanguage.ENGLISH.getLanguageId());
+
+        if (language == null) {
+            throw new IllegalStateException("Fallback language \"en\" is missing!");
+        }
+
+        FileConfiguration fallback = language.getFile(file);
+
+        if (fallback != null) {
+            return fallback;
+        } else {
+            throw new IllegalStateException("Fallback file: \"" + file.getFilePath("en") + "\" is missing!");
+        }
+    }
+
+    @Nonnull
+    @Override
+    public String getMessage(@Nonnull String key) {
+        Validate.notNull(key, "Message key must not be null!");
+
+        Language language = getDefaultLanguage();
+
+        String message = language == null ? null : language.getFile(LanguageFile.MESSAGES).getString(key);
+
+        if (message == null) {
+            return getFallback(LanguageFile.MESSAGES).getString(key);
+        }
+
+        return message;
+    }
+
+    @Nonnull
+    public String getMessage(@Nonnull Player p, @Nonnull String key) {
+        Validate.notNull(p, "Player must not be null!");
+        Validate.notNull(key, "Message key must not be null!");
+
         Language language = getLanguage(p);
 
         if (language == null) {
             return "NO LANGUAGE FOUND";
         }
 
-        String message = language.getMessagesFile().getString(key);
+        String message = language.getFile(LanguageFile.MESSAGES).getString(key);
 
         if (message == null) {
-            Language fallback = getLanguage(SupportedLanguage.ENGLISH.getLanguageId());
-            return fallback.getMessagesFile().getString(key);
+            return getFallback(LanguageFile.MESSAGES).getString(key);
         }
 
         return message;
     }
 
-    public List<String> getMessages(Player p, String key) {
+    @Nonnull
+    public List<String> getMessages(@Nonnull Player p, @Nonnull String key) {
+        Validate.notNull(p, "Player should not be null.");
+        Validate.notNull(key, "Message key cannot be null.");
+
         Language language = getLanguage(p);
 
         if (language == null) {
-            return Arrays.asList("NO LANGUAGE FOUND");
+            return Collections.singletonList("NO LANGUAGE FOUND");
         }
 
-        List<String> messages = language.getMessagesFile().getStringList(key);
+        List<String> messages = language.getFile(LanguageFile.MESSAGES).getStringList(key);
 
         if (messages.isEmpty()) {
-            Language fallback = getLanguage(SupportedLanguage.ENGLISH.getLanguageId());
-            return fallback.getMessagesFile().getStringList(key);
+            return getFallback(LanguageFile.MESSAGES).getStringList(key);
         }
 
         return messages;
     }
 
+    @Nonnull
+    @ParametersAreNonnullByDefault
     public List<String> getMessages(Player p, String key, UnaryOperator<String> function) {
+        Validate.notNull(p, "Player cannot be null.");
+        Validate.notNull(key, "Message key cannot be null.");
+        Validate.notNull(function, "Function cannot be null.");
+
         List<String> messages = getMessages(p, key);
         messages.replaceAll(function);
 
         return messages;
     }
 
-    public String getResearchName(Player p, NamespacedKey key) {
+    @Nullable
+    public String getResearchName(@Nonnull Player p, @Nonnull NamespacedKey key) {
+        Validate.notNull(p, "Player must not be null.");
+        Validate.notNull(key, "NamespacedKey cannot be null.");
+
         Language language = getLanguage(p);
 
-        if (language.getResearchesFile() == null) {
+        if (language == null || language.getFile(LanguageFile.RESEARCHES) == null) {
             return null;
         }
 
-        return language.getResearchesFile().getString(key.getNamespace() + "." + key.getKey());
+        return language.getFile(LanguageFile.RESEARCHES).getString(key.getNamespace() + '.' + key.getKey());
     }
 
-    public String getCategoryName(Player p, NamespacedKey key) {
+    @Nullable
+    public String getCategoryName(@Nonnull Player p, @Nonnull NamespacedKey key) {
+        Validate.notNull(p, "Player must not be null.");
+        Validate.notNull(key, "NamespacedKey cannot be null!");
+
         Language language = getLanguage(p);
 
-        if (language.getCategoriesFile() == null) {
+        if (language == null || language.getFile(LanguageFile.CATEGORIES) == null) {
             return null;
         }
 
-        return language.getCategoriesFile().getString(key.getNamespace() + "." + key.getKey());
+        return language.getFile(LanguageFile.CATEGORIES).getString(key.getNamespace() + '.' + key.getKey());
     }
 
-    public String getResourceString(Player p, String key) {
+    @Nullable
+    public String getResourceString(@Nonnull Player p, @Nonnull String key) {
+        Validate.notNull(p, "Player should not be null!");
+        Validate.notNull(key, "Message key should not be null!");
+
         Language language = getLanguage(p);
 
-        String value = language.getResourcesFile() != null ? language.getResourcesFile().getString(key) : null;
+        String value = language != null && language.getFile(LanguageFile.RESOURCES) != null ? language.getFile(LanguageFile.RESOURCES).getString(key) : null;
 
         if (value != null) {
             return value;
         } else {
-            Language fallback = getLanguage(SupportedLanguage.ENGLISH.getLanguageId());
-            return fallback.getResourcesFile().getString(key);
+            return getFallback(LanguageFile.RESOURCES).getString(key);
         }
     }
 
-    public ItemStack getRecipeTypeItem(Player p, RecipeType recipeType) {
+    @Nullable
+    public ItemStack getRecipeTypeItem(@Nonnull Player p, @Nonnull RecipeType recipeType) {
+        Validate.notNull(p, "Player cannot be null!");
+        Validate.notNull(recipeType, "Recipe type cannot be null!");
+
         Language language = getLanguage(p);
         ItemStack item = recipeType.toItem();
         NamespacedKey key = recipeType.getKey();
 
-        if (language.getRecipeTypesFile() == null || !language.getRecipeTypesFile().contains(key.getNamespace() + "." + key.getKey())) {
+        if (language == null || language.getFile(LanguageFile.RECIPES) == null || !language.getFile(LanguageFile.RECIPES).contains(key.getNamespace() + '.' + key.getKey())) {
             language = getLanguage("en");
         }
 
-        if (!language.getRecipeTypesFile().contains(key.getNamespace() + "." + key.getKey())) {
+        if (!language.getFile(LanguageFile.RECIPES).contains(key.getNamespace() + '.' + key.getKey())) {
             return item;
         }
 
-        FileConfiguration config = language.getRecipeTypesFile();
+        FileConfiguration config = language.getFile(LanguageFile.RECIPES);
 
         return new CustomItem(item, meta -> {
             meta.setDisplayName(ChatColor.AQUA + config.getString(key.getNamespace() + "." + key.getKey() + ".name"));
@@ -218,7 +284,10 @@ public abstract class SlimefunLocalization extends Localization implements Keyed
     }
 
     @Override
-    public void sendMessage(CommandSender recipient, String key, boolean addPrefix) {
+    public void sendMessage(@Nonnull CommandSender recipient, @Nonnull String key, boolean addPrefix) {
+        Validate.notNull(recipient, "Recipient cannot be null!");
+        Validate.notNull(key, "Message key cannot be null!");
+
         String prefix = addPrefix ? getPrefix() : "";
 
         if (recipient instanceof Player) {
@@ -229,6 +298,9 @@ public abstract class SlimefunLocalization extends Localization implements Keyed
     }
 
     public void sendActionbarMessage(@Nonnull Player player, @Nonnull String key, boolean addPrefix) {
+        Validate.notNull(player, "Player cannot be null!");
+        Validate.notNull(key, "Message key cannot be null!");
+
         String prefix = addPrefix ? getPrefix() : "";
         String message = ChatColors.color(prefix + getMessage(player, key));
 
@@ -237,15 +309,17 @@ public abstract class SlimefunLocalization extends Localization implements Keyed
     }
 
     @Override
-    public void sendMessage(CommandSender recipient, String key) {
+    public void sendMessage(@Nonnull CommandSender recipient, @Nonnull String key) {
         sendMessage(recipient, key, true);
     }
 
+    @ParametersAreNonnullByDefault
     public void sendMessage(CommandSender recipient, String key, UnaryOperator<String> function) {
         sendMessage(recipient, key, true, function);
     }
 
     @Override
+    @ParametersAreNonnullByDefault
     public void sendMessage(CommandSender recipient, String key, boolean addPrefix, UnaryOperator<String> function) {
         if (SlimefunPlugin.getMinecraftVersion() == MinecraftVersion.UNIT_TEST) {
             return;
@@ -261,7 +335,7 @@ public abstract class SlimefunLocalization extends Localization implements Keyed
     }
 
     @Override
-    public void sendMessages(CommandSender recipient, String key) {
+    public void sendMessages(@Nonnull CommandSender recipient, @Nonnull String key) {
         String prefix = getPrefix();
 
         if (recipient instanceof Player) {
@@ -278,6 +352,7 @@ public abstract class SlimefunLocalization extends Localization implements Keyed
     }
 
     @Override
+    @ParametersAreNonnullByDefault
     public void sendMessages(CommandSender recipient, String key, boolean addPrefix, UnaryOperator<String> function) {
         String prefix = addPrefix ? getPrefix() : "";
 
@@ -294,8 +369,24 @@ public abstract class SlimefunLocalization extends Localization implements Keyed
         }
     }
 
+    @ParametersAreNonnullByDefault
     public void sendMessages(CommandSender recipient, String key, UnaryOperator<String> function) {
         sendMessages(recipient, key, true, function);
     }
 
+    @Nonnull
+    protected Set<String> getTotalKeys(@Nonnull Language lang) {
+        return getKeys(lang.getFiles());
+    }
+
+    @Nonnull
+    protected Set<String> getKeys(@Nonnull FileConfiguration... files) {
+        Set<String> keys = new HashSet<>();
+
+        for (FileConfiguration cfg : files) {
+            keys.addAll(cfg.getKeys(true));
+        }
+
+        return keys;
+    }
 }