+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/gps/PersonalActivationPlate.java
@@ -1,23 +1,38 @@
 package io.github.thebusybiscuit.slimefun4.implementation.items.gps;
 
+import javax.annotation.Nonnull;
+import javax.annotation.ParametersAreNonnullByDefault;
+
+import org.bukkit.block.Block;
+import org.bukkit.entity.Player;
 import org.bukkit.event.block.BlockPlaceEvent;
 import org.bukkit.inventory.ItemStack;
 
 import io.github.thebusybiscuit.slimefun4.core.handlers.BlockPlaceHandler;
 import me.mrCookieSlime.Slimefun.Lists.RecipeType;
 import me.mrCookieSlime.Slimefun.Objects.Category;
-import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SlimefunItem;
 import me.mrCookieSlime.Slimefun.api.BlockStorage;
 import me.mrCookieSlime.Slimefun.api.SlimefunItemStack;
 
-public class PersonalActivationPlate extends SlimefunItem {
+/**
+ * The {@link PersonalActivationPlate} is a teleporter activation plate
+ * to which only the {@link Player} who placed it down has access.
+ * 
+ * @author TheBusyBiscuit
+ * 
+ * @see SharedActivationPlate
+ *
+ */
+public class PersonalActivationPlate extends AbstractTeleporterPlate {
 
+    @ParametersAreNonnullByDefault
     public PersonalActivationPlate(Category category, SlimefunItemStack item, RecipeType recipeType, ItemStack[] recipe) {
         super(category, item, recipeType, recipe);
 
         addItemHandler(onPlace());
     }
 
+    @Nonnull
     private BlockPlaceHandler onPlace() {
         return new BlockPlaceHandler(false) {
 
@@ -26,5 +41,11 @@ public class PersonalActivationPlate extends SlimefunItem {
                 BlockStorage.addBlockInfo(e.getBlock(), "owner", e.getPlayer().getUniqueId().toString());
             }
         };
+    }
+
+    @Override
+    @ParametersAreNonnullByDefault
+    public boolean hasAccess(Player p, Block b) {
+        return BlockStorage.getLocationInfo(b.getLocation(), "owner").equals(p.getUniqueId().toString());
     }
 }