+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Table of contents
-- [Release Candidate 22 (TBD)](#release-candidate-22-tbd)
+- [Release Candidate 23 (TBD)](#release-candidate-23-tbd)
+- [Release Candidate 22 (18 Apr 2021)](#release-candidate-22-18-apr-2021)
 - [Release Candidate 21 (14 Mar 2021)](#release-candidate-21-14-mar-2021)
 - [Release Candidate 20 (30 Jan 2021)](#release-candidate-20-30-jan-2021)
 - [Release Candidate 19 (11 Jan 2021)](#release-candidate-19-11-jan-2021)
@@ -22,7 +23,30 @@
 - [Release Candidate 2 (29 Sep 2019)](#release-candidate-2-29-sep-2019)
 - [Release Candidate 1 (26 Sep 2019)](#release-candidate-1-26-sep-2019)
 
-## Release Candidate 22 (TBD)
+## Release Candidate 23 (TBD)
+
+#### Additions
+* Added "Quartz Block -> 4 Quartz" recipe to Grind Stone
+* Added "8 Coal Blocks -> 9 Carbon" recipe to Compressor
+* Added "8 Coal Blocks -> 9 Carbon" recipe to Carbon Press
+* Added Tier 2 Auto-Enchanter
+* Added Tier 2 Auto-Disenchanter
+* (API) Added Category#setTier() to modify a category's position in the guide
+* Added the ability to disable auto (dis)enchanting with a lore - `use-ignored-lores` & `ignored-lores` in Items.yml
+
+#### Changes
+* Renamed "Solar Panel" to "Photovoltaic Cell" to avoid confusions with solar generators
+* Photovoltaic Cells can no longer be placed
+* (API) Removed deprecated "SlimefunBlockHandler"
+* Removed Automated Crafting Chamber
+
+#### Fixes
+* Fixed #2987
+* Fixed #2989
+* Fixed #2977
+
+## Release Candidate 22 (18 Apr 2021)
+https://thebusybiscuit.github.io/builds/TheBusyBiscuit/Slimefun4/stable/#22
 
 #### Additions
 * Added Vanilla Auto-Crafter
@@ -40,6 +64,7 @@
 * Added "magma cream -> slime ball" recipe to the Freezer
 * Added "2 magma blocks -> slime block" recipe to the Freezer
 * Added configurable enchantment level limit for both auto enchanter and auto disenchanter
+* (API) Added AutoEnchantEvent
 
 #### Changes
 * Changed item order in guide for the Villager Rune and Nether Goo (All runes are now grouped together)
@@ -69,6 +94,8 @@
 * Fixed #2926
 * Fixed Grappling Hook vanishing in creative mode
 * Fixed #2944
+* Fixed #2837
+* Fixed #2942
 
 ## Release Candidate 21 (14 Mar 2021)
 https://thebusybiscuit.github.io/builds/TheBusyBiscuit/Slimefun4/stable/#21